Terrier::Embedded: move field methods to class methods
the distinction is that when you use `included do`, it will define methods directly on th e class that includes Embedded, rather than defining them as class methods on the module that get inherited. Defining them as class methods on Embedded makes the chain of inheritance cleaner and allows for subclasses of Embedded to override the behavior nicely.